requirements.txt
boto3==1.9.66
botocore==1.12.66
certifi==2018.11.29
chardet==3.0.4
click==7.0
docutils==0.14
flask==1.0.2
idna==2.8
itsdangerous==1.1.0
jinja2==2.10
jmespath==0.9.3
markupsafe==1.1.0
python-dateutil==2.7.5; python_version >= '2.7'
pyyaml==3.13
requests==2.21.0
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1; python_version >= '3.4'
werkzeug==0.14.1
